+++ components/src/lib/Modal/BackDrop/style.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const BackDrop = styled.div`
+  position: fixed;
+  top: 0;
+  right: 0;
+  bottom: 0;
+  left: 0;
+
+  background-color: rgba(0, 0, 0, 0.35);
+`;
