+++ components/src/lib/Modal/CloseButton/style.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const CloseButton = styled.img`
+  width: 14px;
+  height: 14px;
+
+  margin: 0 10px;
+`;
