+++ components/src/lib/Modal/Header/style.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+export const Header = styled.div`
+  height: 26px;
+
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+`;
