+++ components/src/lib/Modal/Title/style.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const Title = styled.h1`
+  color: #000000;
+  font-size: 18px;
+  font-weight: 700;
+  line-height: 26px;
+`;
