+++ components/src/lib/Modal/style.tsx
@@ -1,0 +1,8 @@
+import styled from "styled-components";
+
+export const Modal = styled.section`
+  position: relative;
+
+  height: 100vh;
+  width: 376px;
+`;
